Revert "Revert "Vision API features update (#1339)" (#1351)"
This reverts commit 7b39e19557ca5bdfe323cc6ca95ce2fa1303208f.
